[#42296389] Remove rbenv-alternatives
We don't actually have a need for this, as we only ever want to use a single
system Ruby binary. The patch also isn't in this upstream version of rbenv
so removing it simplifies our install.

NB: There was a permissions issue which I fixed with `override_dh_fixperms`
to set execute bits on the rbenv-alternatives binary. If we reinstate this
with a patch again then we'll need to do that.