Rename directory out of tarball
Ensures that it matches debuild's expectation of `name-version`. Deletes any
existing copies before extracting/renaming.

Previously the DSC was referencing a codebase that only contained our
`debian/` directory because the tarball contained a different name.

This probably won't handle tarballs that don't contain a parent directory. I
expect that something bad will happen. We'll tackle that when it arises.