Take directory instead of name as argument
Makes it easy to auto-complete on the CLI. Does away with the need for
`self.name`. The build directory has been renamed to match `name-version`,
per `expanded_dir`, because it's easier to construct this from the details
in the changelog. This will also prevent conflicts where we have multiple
versions of package sources in our Git repo (eg. `ruby1.9.2` actually
creates a package called `ruby`).