[#42296389] Initial manifests for rbenv-ruby-1.9.3
Installs Ruby to a prefix directory. The parent of which is owned by the
rbenv package, which is listed as a dependency.

Package name contains the entire version string so that we can have multiple
unique versions (e.g. 1.9.2-p290, 1.9.2-p320, 1.9.3-p392) in the same repo.

This version of Ruby ships with rubygems 1.8.23 and thus doesn't need an
updated version.