Include vegeta dependency as a patch
This is a pretty horrible workaround for:

a) Launchpad build farm doesn't allow outbound connectivity during builds.
   So `go get` fails.
b) Precise doesn't support `dh-golang` which automates some of this for us.

Stuff the only dependency into a patch and apply it during build time. I've
intentionally included the `.git` directory so that we have some record of
WTH version it is.