Setup toolchain to build on a trusty VM
Also update docs to refer to Precise and Trusty as opposed to Lucid and
Precise.

This will make the packager VM run trusty instead of precise, so that
the more recent tools will be available to build_source.py.  The actual
builds still happen with a schroot for the appropriate distro (or at
launchpad...).