Update godep package to build on trusty.
Trusty ships with a recent enough version of golang, so it's easiest to
build here.  The resulting binary package can then be used on both
trusty and precise because go binaries are very portable.

Note: the upstream godep project does not have any versioning at the
moment, so I've had to invent a version.  This has been generated by
locally tagging the first commit in the repository, and then using `git
describe --tags`