Update versioning for godep.
The upstream author has now created a tag for the first commit in the
repo[1].  This means we can now use version numbers as generated by `git
describe`.  This has resulted in a decrease in the version number.  This
hasn't been widely deployed as yet, so can be cleaned up manually.

[1] https://github.com/tools/godep/issues/126