Package locksmithctl version 0.1.8
Package `locksmithctl(1)`, a simple client to control a lock (or mutex)
in an etcd backend, which we can use to control machine reboots to
ensure only a specified number of machines is rebooted at any one time.

See also:
https://github.com/coreos/locksmith#usage

This build is for Trusty, but since it's a Go binary with almost no
dependencies, we can use the same binary for Precise.