Remove default etcd configuration
Remove the package-supplied etcd.conf to prevent a race condition
whereby an etcd cluster is bootstrapped using the supplied etcd.conf and
the service started before we have a chance to modify etcd.conf (using
Puppet, for example).

If the etcd cluster comes up using the default configuration, then we
change that configuration, the nodes of the cluster will continue trying
to connect to the peers listed in the default configuration.

By omitting the default configuration, the etcd service will fail to
start until Puppet has placed the correct configuration in
`/etc/etcd/etcd.conf`.