Upgrade Gor to version 0.10.1
There are a lot of bug fixes since 0.7.X as well
new features. One of the new features we need is
the ability to disallow certain url patterns using
a regex.

As this is a GO binary, there is no need to build
seperate packages for precise/trusty. The same
package should work on either release.

This package have successfully uploaded to the GOV.UK PPA
and I have tested installing OK on a Vagrant box.