Create packages for hiera-eyaml
This creates the packages for both hiera-eyaml and hiera-eyaml-gpg. We
previously packaged these up to go on PPA, but did not have them
available for a Xenial release. This uses the built in fpm-cookery
function to create a package directly from rubygems. The end result are
packages called "rubygem-hiera-eyaml" and "rubygem-hiera-eyaml-gpg"
respectively.

fpm functionality described here:
https://github.com/jordansissel/fpm/wiki/ConvertingGems