Replace the RubyGemRecipe with a custom one.
Xenial does not provide a package for hiera-eyaml-gpg which we use for
encrypting hiera files so we have to make our own package.

fpm-cookery does provide a 'RubyGemRecipe' but it mimics whichever install
path the enviroment uses for gems. This causes problems when e.g. rbenv is
being used as the package will install gems to directories that aren't in
PATH. To make sure we always use the standard install location we have hard
coded the paths in this recipe.