Add logstash 1.1.9
Add recipe to package the Logstash jar file. The file is quite big,
so we are not going to include it in this repo. To create the package,
download the jar file from `https://logstash.objects.dreamhost.com/release/logstash-1.1.9-monolithic.jar`
to `/var/tmp/`.

This recipe won't run on Jenkins, run it locally and copy the file to
the Aptly machine.