Add Vale lint package
ValeLint is a super userful tool to run against any documentation as you
can configure it to check your documentation and produce warnings or
errors to your specific standards.

This creates a debian package to allow us to run it in our CI
environment against the GOV.UK developer docs.