Invalidate DST_Root_CA_X3 CA cert when using trusty image
As per https://ubuntu.com/security/notices/USN-5089-2
We invalidate this expired CA.
The proper way to do this would be to update the ca-certificates
package. However trusty is not supported anymore and this update
would have to be done through an authenticated ESM repo.
We therefore do this manually in order to minimize the efforts
required.